chore: delete old files
